+++ pynenc_tests/unit/orchestrator/test_cycle_control_all_instances.py
@@ -73,10 +73,10 @@ def test_causes_cycle(invocations: tuple, app_instance: "Pynenc") -> None:
 
     expected_error = (
         "A cycle was detected: Cycle detected:\n"
-        "- Call(task=pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task2, arguments=<no_args>)\n"
-        "- Call(task=pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task0, arguments=<no_args>)\n"
-        "- Call(task=pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task1, arguments=<no_args>)\n"
-        "- back to Call(task=pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task2, arguments=<no_args>)"
+        "- #task_id#pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task2#args_id#no_args\n"
+        "- #task_id#pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task0#args_id#no_args\n"
+        "- #task_id#pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task1#args_id#no_args\n"
+        "- back to #task_id#pynenc_tests.unit.orchestrator.test_cycle_control_all_instances.task2#args_id#no_args"
     )
 
     assert str(exc_info.value) == expected_error